Proyecto/requirements.txt
tensorflow
tensorflow-datasets
requests
imdbpy